+++ src/ui/BaseContextMenu.js
@@ -68,7 +68,7 @@ export class BaseContextMenu {
         this.isMonitoringCursor = false;
         this.animationStartTime = 0;
 
-        console.log('BaseContextMenu: Constructor called for panel', panel);
+        Logger.ui.debug('BaseContextMenu: Constructor called for panel', panel);
         this.setupContextMenu();
         
         // Only setup global handlers if not disabled
@@ -228,7 +228,7 @@ export class BaseContextMenu {
     setupContextMenu() {
         // Remove existing context menu handler if it exists
         if (this.contextMenuHandler) {
-            console.log('BaseContextMenu: Removing existing context menu handler from', this.panel);
+            Logger.ui.debug('BaseContextMenu: Removing existing context menu handler from', this.panel);
             this.panel.removeEventListener('contextmenu', this.contextMenuHandler);
         }
         
@@ -242,7 +242,7 @@ export class BaseContextMenu {
             this.showContextMenu(e, contextData);
         };
 
-        console.log('BaseContextMenu: Adding context menu handler to', this.panel);
+        Logger.ui.debug('BaseContextMenu: Adding context menu handler to', this.panel);
         this.panel.addEventListener('contextmenu', this.contextMenuHandler);
     }
 
@@ -926,13 +926,13 @@ export class BaseContextMenu {
      * Destroy context menu and clean up event listeners
      */
     destroy() {
-        console.log('BaseContextMenu: Destroying context menu for', this.panel);
+        Logger.ui.debug('BaseContextMenu: Destroying context menu for', this.panel);
         this.hideMenu();
         if (this.resizeHandler) {
             window.removeEventListener('resize', this.resizeHandler);
         }
         if (this.contextMenuHandler) {
-            console.log('BaseContextMenu: Removing context menu handler from', this.panel);
+            Logger.ui.debug('BaseContextMenu: Removing context menu handler from', this.panel);
             this.panel.removeEventListener('contextmenu', this.contextMenuHandler);
         }
         this.removeCursorTracking();